% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniq.col.value.R
\name{uniq.col.value}
\alias{uniq.col.value}
\title{uniq.col.value
take a data frame and return how many unique values each columns has}
\usage{
uniq.col.value(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
Numbers of unique values in each column
}
\description{
uniq.col.value
take a data frame and return how many unique values each columns has
}
\examples{
data(mtcars) # from datasets-package {datasets}
uniq.col.value(mtcars)

data(iris) # datasets-package {datasets}
uniq.col.value(iris)
}
\references{
https://discuss.analyticsvidhya.com/t/how-to-count-number-of-distinct-values-in-a-column-of-a-data-table-in-r/1124
}
