% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniq.row.value.R
\name{uniq.row.value}
\alias{uniq.row.value}
\title{uniq.row.value}
\usage{
uniq.row.value(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
Numbers of unique values in each row
}
\description{
take a data frame and return how many unique values each row has
}
\examples{

data(mtcars) # from datasets-package {datasets}
uniq.row.value(mtcars)

data(iris) # datasets-package {datasets}
uniq.row.value(iris)
}
\references{
https://www.tutorialspoint.com/how-to-find-the-number-of-unique-values-in-each-row-of-an-r-data-frame
}
