% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniquerYC.R
\docType{package}
\name{uniquerYC-package}
\alias{uniquerYC}
\alias{uniquerYC-package}
\title{uniquerYC}
\arguments{
\item{\cr}{\itemize{
\item uniq.col.value() - take a data frame and return how many unique values each columns has
\item uniq.row.value() - take a data frame and return how many unique values each row has
}
search ?uniq.col.value() or ?uniq.row.value() to get more information on how to use these functions.}
}
\description{
the uniquerYC package provides functions
}
\section{UniquerYC has two functions}{

uniq.col.value() and uniq.row.value() \cr
}

\author{
\strong{Maintainer}: Yuka Chen \email{yjchen9596@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
